+++ src/message_interface/message_handler.rs
@@ -976,7 +976,7 @@ impl WalletMessageHandler {
                 .await
             }
             AccountMethod::RequestFundsFromFaucet {
-                faucet_url: url,
+                url: url,
                 address,
             } => {
                 convert_async_panics(|| async {