+++ bindings/java/CHANGELOG.md
@@ -32,10 +32,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Fix `RequiredStorageDeposit` deserialization;
 - Rename `ClientCommand` to `WalletCommand`;
 - Move `WalletCommand` class from `NativeApi`;
 - Rename `ClientResponse` to `WalletResponse`;
+
+### Fixed
+
+- `RequiredStorageDeposit` deserialization;
 
 <!--
 ## 1.0.0-rc.2 - 2022-11-15