+++ bindings/java/CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
-## 1.0.0-rc.1 - 2022-MM-DD
+## 1.0.0-rc.1 - 2022-12-06
 
 ### Added